requirements.txt
Telethon
pandas
openpyxl
google-generativeai
python-dotenv
